+++ semcore/mini-chart/CHANGELOG.md
@@ -12,6 +12,12 @@ CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangel
 
 - `ScoreDonut` and `ScoreSemiDonut` visual value can't be greater than 100.
 
+## [16.0.9] - 2025-09-17
+
+### Changed
+
+- Version patch update due to children dependencies update (`@semcore/base-components` [16.2.2 ~> 16.2.3], `@semcore/flex-box` [16.0.8 ~> 16.0.9]).
+
 ## [16.0.8] - 2025-09-05
 
 ### Changed